+++ src/lib/modules/vars.ts
@@ -4,10 +4,6 @@ export const slugPost = 'blog'
 
 export const slugPortofolio = 'portofolios'
 
-export const slugTool = 'stacks'
-
-export const slugCertification = 'certifications'
-
 export const slugReusable = 'reusable'
 
 export const slug404 = '404'