+++ src/ui/layouts/Certification.tsx
@@ -84,7 +84,7 @@ export function CertificationCard({ data, ...props }: CertificationCardProps) {
 								<Badge
 									key={`${compId}-skill-${index}`}
 									variant="light"
-									color="dark"
+									color={themeColor === 'light' ? 'dark' : 'gray'}
 								>
 									{skill.title}
 								</Badge>